Modification of the cardiality of resource data from 1-1 to 1-n.
This includes multiple meta_data rows per object in the sqlite3 database.  Also the passing of the MediaResource in the APIs that assume a 1-1 relationship between item and resource data.  Container and non-video item media objects are currently stubbed out implementations, and will require more investigation.

Also all non-resource based data should be moved to object from meta_data as it is now duplicated.  Including class, data, and creator.

Noticed regression with gupnp-av-control which queries the same container in an infinate loop.  This looks to have been introduced earlier and just noticed here.