Made some changes to tighten up the code prior to generation of Patch A.
Changed ConnectionManagerProtocolInfo.update_source_protocol_info() to
 skip MediaItems with unset media_resource lists, preventing glib assertion
 errors.
Changed MediaItem.write_didl_lite_for_resources() to create a URI for
 resources if the object's uri is null or empty. It's now set empty instead
 of null by MediaExport to prevent assertion errors.
Demoted a bunch of message-level logging to debug-level in HTTPGet,
 ODIDMediaEngine, HTTPMediaResourceHandler.
Added TODOs everywhere that we expect to change in the upcoming patches.
Added a missing null check for seek_request. This allows seek-less streaming
 requests to be processed.

Tested Range and non-Range requests with the Simple and Gst MediaEngines
 with curl and VLC.